+++ packages/phishing-controller/src/types.ts
@@ -198,6 +198,20 @@ export const DEFAULT_CHAIN_ID_TO_NAME = {
   '0x82': 'unichain',
   '0x7e4': 'ronin',
   '0x127': 'hedera',
+  '0x12c': 'zksync-sepolia',
+  '0xaa36a7': 'ethereum-sepolia',
+  '0xa869': 'avalanche-fuji',
+  '0x343b': 'immutable-zkevm',
+  '0x34a1': 'immutable-zkevm-testnet',
+  '0x64': 'gnosis',
+  '0x1e0': 'worldchain',
+  '0x8173': 'apechain',
+  '0x138c5': 'berachain-bartio',
+  '0xdef1': 'ink',
+  '0xba5ed': 'ink-sepolia',
+  '0x2b74': 'abstract-testnet',
+  '0x531': 'sei',
+  '0x2eb': 'flow-evm',
 } as const;
 
 export type ChainIdToNameMap = typeof DEFAULT_CHAIN_ID_TO_NAME;
@@ -218,6 +232,10 @@ export enum AddressScanResultType {
    * Address is malicious
    */
   Malicious = 'Malicious',
+  /**
+   * Error occurred during scan
+   */
+  ErrorResult = 'ErrorResult',
 }
 
 /**