+++ packages/phishing-controller/src/index.ts
@@ -7,6 +7,10 @@ export type {
   PhishingDetectorConfiguration,
 } from './PhishingDetector';
 export { PhishingDetector } from './PhishingDetector';
-export type { PhishingDetectionScanResult } from './types';
-export { PhishingDetectorResultType, RecommendedAction } from './types';
+export type { PhishingDetectionScanResult, AddressScanResult } from './types';
+export {
+  PhishingDetectorResultType,
+  RecommendedAction,
+  AddressScanResultType,
+} from './types';
 export type { CacheEntry } from './CacheManager';