+++ packages/eth-json-rpc-middleware/src/inflight-cache.ts
@@ -11,7 +11,10 @@ import {
 import { projectLogger, createModuleLogger } from './logging-utils';
 import { cacheIdentifierForRequest } from './utils/cache';
 
-type RequestHandler = [(result: Json) => void, (error: Error) => void];
+type RequestHandler = [
+  (result: Readonly<Json>) => void,
+  (error: unknown) => void,
+];
 
 type InflightRequest = {
   [cacheId: string]: RequestHandler[];
@@ -64,7 +67,7 @@ export function createInflightCacheMiddleware(): JsonRpcMiddleware<
     // allow request to be handled normally
     log('Carrying original request forward %o', request);
     try {
-      const result = (await next()) as Json;
+      const result = (await next()) as Readonly<Json>;
       log(
         'Running %i collected handler(s) for successful request %o',
         activeRequestHandlers.length,
@@ -78,7 +81,7 @@ export function createInflightCacheMiddleware(): JsonRpcMiddleware<
         activeRequestHandlers.length,
         request,
       );
-      runRequestHandlers({ error: error as Error }, activeRequestHandlers);
+      runRequestHandlers({ error }, activeRequestHandlers);
       throw error;
     } finally {
       delete inflightRequests[cacheId];
@@ -94,11 +97,11 @@ export function createInflightCacheMiddleware(): JsonRpcMiddleware<
  */
 function createActiveRequestHandler(
   activeRequestHandlers: RequestHandler[],
-): Promise<Json> {
-  const { resolve, promise, reject } = createDeferredPromise<Json>();
+): Promise<Readonly<Json>> {
+  const { resolve, promise, reject } = createDeferredPromise<Readonly<Json>>();
   activeRequestHandlers.push([
-    (result: Json) => resolve(result),
-    (error: Error) => reject(error),
+    (result: Readonly<Json>) => resolve(result),
+    (error: unknown) => reject(error),
   ]);
   return promise;
 }
@@ -110,7 +113,7 @@ function createActiveRequestHandler(
  * @param activeRequestHandlers - The active request handlers.
  */
 function runRequestHandlers(
-  resultOrError: { result: Json } | { error: Error },
+  resultOrError: { result: Readonly<Json> } | { error: unknown },
   activeRequestHandlers: RequestHandler[],
 ): void {
   // use setTimeout so we can handle the original request first