+++ packages/notification-services-controller/src/NotificationServicesPushController/services/endpoints.ts
@@ -1,2 +1,13 @@
-const url = 'https://push.api.cx.metamask.io';
-export const REGISTRATION_TOKENS_ENDPOINT = `${url}/api/v2/token`;
+export type ENV = 'prd' | 'uat' | 'dev';
+
+const PUSH_API_ENV = {
+  dev: 'https://push.dev-api.cx.metamask.io',
+  uat: 'https://push.uat-api.cx.metamask.io',
+  prd: 'https://push.api.cx.metamask.io',
+} satisfies Record<ENV, string>;
+
+export const PUSH_API = (env: ENV = 'prd') =>
+  PUSH_API_ENV[env] ?? PUSH_API_ENV.prd;
+
+export const REGISTRATION_TOKENS_ENDPOINT = (env: ENV = 'prd') =>
+  `${PUSH_API(env)}/api/v2/token`;