+++ packages/phishing-controller/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add address scanning to detect malicious addresses ([#7118](https://github.com/MetaMask/core/pull/7118))
   - Add `scanAddress` method to scan addresses for security alerts
-  - Add `AddressScanResult` and `AddressScanCacheData` types
+  - Add `AddressScanResult` type
   - Add `addressScanCache` to `PhishingControllerState`
   - Add action registration for `scanAddress` method as `PhishingControllerScanAddressAction`